Set default http agent max sockets
Effectively, this sets maximum concurrent requests. The http connection pool
will not grow larger than the given number. When the limit is reached, requests
are internally queued, so no special handling in the client app is necessary.

The maxTotalSockets doesn't really matter in our case, where the agent is used
for requests to a single host, but it's there for safety.